+++ changelog.md
@@ -2,6 +2,7 @@
 - Remove VALVE_STATE from HmIPW-FALMOT-C12
 - Remove support for python 3.9
 - Remove old-style union syntax
+- Remove to int converter for HmIP-SCTH230 CONCENTRATION
 
 # Version 2023.1.7 (2023-01-24)
 - Aggregate calls to backend